♻ [Refactor] Move some FlatPageForm tests
